Documentation refresh
This commit is confusing because I accidentally recreated this
repo in another repo as I couldn't locate it.

In that repo I created version 0.0.6 and pushed that version to
RubyGems. I didn't realize that 0.0.6 included a crashing bug so I
had to yank it. This is why this commit contains a bump to 0.0.7,
in order to avoid any accidental re-release of a 0.0.6 version.

I will trash the other repo I created and rename this one from
oa-codepath (which doesn't match the gem name anymore) to
oa-codeschool (which does).